TH/requirements.txt
pandas
pyyaml
python-Levenshtein
scikit-learn
textblob


